Move symbols into Transform
WindowProperty
